requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fasttext==0.9.3
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.10.0
gevent==23.9.1
greenlet==3.1.1
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.23.5
packaging==24.2
pandas==2.2.3
pybind11==2.13.6
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
six==1.16.0
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.1.1
